Rename SurveyFooter to ZUIPublicFooter
